requirements.txt
requests
beautifulsoup4
mysql-connector-python
python-dotenv
selenium
webdriver-manager
flask
flask-cors